requirements.txt
pandas==1.5.3
scikit-learn==1.1.1
numpy==1.23.5
mlflow-skinny==2.9.2
category-encoders==2.6.3
pre-commit==3.3.2
pylint=2.11.1
black==23.3.0